requirements.txt
Django~=4.2.5
dropbox~=11.36.2
cloudinary~=1.34.0
requests~=2.31.0
python-dotenv~=1.0.0
phonenumbers~=8.13.20